+++ src/shapes/Transformer.js
@@ -12,7 +12,8 @@
     'borderStrokeWidthChange',
     'anchorStrokeChange',
     'anchorStrokeWidthChange',
-    'anchorFillChange'
+    'anchorFillChange',
+    'anchorCornerRadiusChange'
   ].join(' ');
 
   var NODE_RECT = 'nodeRect';
@@ -123,6 +124,7 @@
    * @param {Array} [config.borderDash] Array for border dash.
    * @param {String} [config.anchorFill] Anchor fill color
    * @param {String} [config.anchorStroke] Anchor stroke color
+   * @param {String} [config.anchorCornerRadius] Anchor corner radius
    * @param {Number} [config.anchorStrokeWidth] Anchor stroke size
    * @param {Number} [config.anchorSize] Default is 10
    * @param {Boolean} [config.keepRatio] Should we keep ratio when we are moving edges? Default is true
@@ -712,7 +714,8 @@
         offsetY: anchorSize / 2,
         stroke: this.getAnchorStroke(),
         strokeWidth: this.getAnchorStrokeWidth(),
-        fill: this.getAnchorFill()
+        fill: this.getAnchorFill(),
+        cornerRadius: this.getAnchorCornerRadius()
       });
 
       this.findOne('.top-left').setAttrs({
@@ -1025,6 +1028,27 @@
    * transformer.anchorFill('red');
    */
   Konva.Factory.addGetterSetter(Konva.Transformer, 'anchorFill', 'white');
+
+  /**
+   * get/set anchor corner radius
+   * @name anchorCornerRadius
+   * @method
+   * @memberof Konva.Transformer.prototype
+   * @param {Number} enabled
+   * @returns {Number}
+   * @example
+   * // get
+   * var anchorCornerRadius = transformer.anchorCornerRadius();
+   *
+   * // set
+   * transformer.anchorCornerRadius(3);
+   */
+  Konva.Factory.addGetterSetter(
+    Konva.Transformer,
+    'anchorCornerRadius',
+    0,
+    Konva.Validators.getNumberValidator()
+  );
 
   /**
    * get/set border stroke color